remove task sorter struct, rename main.go to tasksorter.go still main package
