+++ cmd/tasksort/tasksort.go
@@ -4,8 +4,6 @@ import (
 	"encoding/json"
 	"log"
 	"net/http"
-	"task-handler/Models"
-	"task-handler/TaskProcessor"
 )
 
 func main() {
